backend/requirements.txt
flask==3.0.3
pypdf==4.0.2
flask_cors==5.0.0
langchain==0.2.16
pinecone-client==5.0.1
python-dotenv==1.0.1
langchain-pinecone==0.1.3
langchain-openai==0.1.25
langchain_community==0.2.17